# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aryan Landmark Solutions API
  version: 1.0.0
  description: |
    Production REST API for Aryan Landmark Solutions.

    PostgreSQL contains six tables. All property photos, property documents
    and inquiry call recordings use one unified `attachments` resource.

    Cloudflare R2 stores file binaries. PostgreSQL stores metadata and the R2
    object key.
servers:
  - url: /api/v1
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login ADMIN or EMPLOYEE
      description: |
        `username` is users.user_id.

        CUSTOMER and MASTER cannot use this endpoint.

        Phase 36: rate-limited per source IP via the LOGIN_RATE_LIMITER
        Cloudflare binding, independent of MASTER login's own limiter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /master-auth/login:
    post:
      tags: [Master Authentication]
      summary: Login Master Record Management
      description: |
        MASTER is manually provisioned in the users table.
        There is no public master-registration endpoint.

        Phase 36: rate-limited per source IP via the
        MASTER_LOGIN_RATE_LIMITER Cloudflare binding, independent of
        application login's own limiter and configured more strictly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Master login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterLoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    applicationBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT

    masterBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: >-
        Phase 36. Too many login attempts from this source IP. Applies
        only to POST /auth/login and POST /master-auth/login, each
        backed by its own independent Cloudflare Rate Limiting binding
        (application login and MASTER login can never exhaust each
        other's budget). The response never reveals which email/account
        was attempted or any internal limiter state (current count,
        reset time) — same generic-failure principle as
        AUTH_INVALID_CREDENTIALS.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: EMP001
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          required: [accessToken, user]
          properties:
            accessToken:
              type: string
            tokenType:
              type: string
              example: Bearer
            expiresIn:
              type: integer
              example: 2592000
              description: >-
                Phase 35: application session lifetime, in seconds (30
                days — raised from the original 3600s). There is no
                refresh token; a session instead ends via explicit
                logout or ADMIN deactivation (enforced server-side on
                the next authenticated request, per requireApplicationAuth's
                live status check), whichever comes first.
            user:
              $ref: '#/components/schemas/User'
    MasterLoginResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          required: [accessToken, tokenScope, isMaster]
          properties:
            accessToken:
              type: string
            tokenType:
              type: string
              example: Bearer
            tokenScope:
              type: string
              example: MASTER
            isMaster:
              type: boolean
              example: true
              description: >-
                Phase 35: explicit capability flag so the client can
                identify a MASTER session directly from the login
                response, instead of inferring it from which login
                endpoint was called. Always true on this response;
                never present on POST /auth/login's response.
            expiresIn:
              type: integer
              example: 3600
              description: MASTER token lifetime is unchanged by Phase 35.
...
    User:
      type: object
      required: [id, userType, name, status]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          nullable: true
        userType:
          type: string
          enum: [APPLICATION_USER, CUSTOMER, MASTER]
        role:
          type: string
          nullable: true
          enum: [ADMIN, EMPLOYEE]
        name:
          type: string
        email:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, INACTIVE, BLOCKED]
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: true
        requestId:
          type: string
...